rename xr_interface module to drivers (simulation drivers)
